Add subproject to represent git-submodule
Git submodules are full git repositories, and so I believe a proper way
to support them in repo is to treat them as if they are just projects.

So instead of adding calls of `git submodule` to every subcmds, we add
new manifest entry "<subproject .../>" to represent git-submodules in
manifest files.

A subproject entry references to a previous-declared project entry and
sets its worktree and relpath to be under parent project's worktree and
relpath directory.

A drawback of this approach is that repo does not keep track of the
git's submodule information, such as the .gitmodules file or the gitlink
tree objects.  As a consequence, you still have to manage submodules
with the `git submodule` command sometimes, for example, when creating
new subprojects.

Note:  I exploit the fact that Command.GetProjects() sorts projects by
relpath alphabetically to make sure that subprojects are checked-out
after their parent project.

Change-Id: I541e9e2ac1a70304272dbe09724572aa1004eb5c