Represent git-submodule as nested projects
We need a representation of git-submodule in repo; otherwise repo will
not sync submodules, and leave workspace in a broken state.  Of course
this would not happen if all projects are owned by the owner of the
manifest file, who may simply choose not to use git-submodule in all
projects and save him the trouble.  However, this is simply not possible
in practice, where some of projects would be downstream repositories
that manifest file owner do not own.

As git submodules are simply git repositories, it is natural to treat
them as plain repo projects that live inside a repo project.  That is,
we could use recursively declared projects to denote the is-submodule
relation of git repositories.

The behavior of repo remains the same to projects that do not have a
sub-project within.  As for parent projects, repo fetches them and their
sub-projects as normal projects, and then checks out subprojects at the
commit specified in parent's commit object.

Note that you have to register a submodule twice (once in repo manifest
and once in the parent Git repository) as repo does not parse .gitmodules
file to collect the registered list of submodules, because .gitmodules
file could be changed through out the history and so relying on it makes
repo unable to have a stable list of projects; imagine how painful it
would be if every time before you run `repo sync` you will have to
manually check out .gitmodules file of parent projects (or make sure
they are checked out).

To a user, a sub-project is merely a git-submodule; so all tips of
working with a git-submodule apply here, too.  For example, you should
not run `repo sync` in a parent repository if its submodule is dirty.

Change-Id: I541e9e2ac1a70304272dbe09724572aa1004eb5c