Represent git-submodule as nested projects
repo uses the notion of "project" to denote a git repository, and
manages a collection of projects.  But repo does not have a notion that
maps to a git submodule.

As git submodules are simply git repositories, it is natural to treat
them as plain repo projects that live inside a repo project.  That is,
we could use recursively declared projects to denote the is-submodule
relation of git repositories.

This patch does not provide automatic management of git-submodule, such
as creating the .gitmodules file.  You have to manage submodules with
the `git submodule` command manually.

Change-Id: I541e9e2ac1a70304272dbe09724572aa1004eb5c