Add 'rebase="false"' attribute to the <project/> XML.
This can prevent 'repo sync' from automatically rebasing.

I hit a situation in where one of the git repositories I was tracking was actually an external repository that I wanted to pull commits into and merge myself. (NOT rebase, since that would lose the merge history.) In this case, I'm not using 'repo upload', I'm manually managing the merges to and from this repository.

Everything was going great until I typed 'repo sync' and it rebased my manually-merged tree. Hence the option to skip it.

Change-Id: I965e0dd1acb87f4a56752ebedc7e2de1c502dbf8