Refine groups functionality
If a project has no groups specified, it is automatically in group
"default".  To specify a group as not default, use groups="".
All group names specified in the manifest are positive names as
opposed to a mix of negative and positive.

Specified groups are resolved in order.  If init is supplied with
--groups="-group1,group2", the following describes the project selection
when syncing:
* all projects in "default" will be added,
* all projects in "group1" will be removed, and
* all projects in "group2" will be selected.

Change-Id: I1df3dcdb64bbd4cd80d675f9b2d3becbf721f661