Update patch set 5
Patch Set 5:

It is possible.  I can check for an optional "keep" attribute.  If it exists and is "true" or "false" (regardless of case), then I can add it to the _Annotation object and use that information to determine whether or not it gets written back.  If it does not exist, I can assume "true".  If it exists and is neither "true" nor "false", I can raise a ManifestParseError with a suitable message.

Out of curiosity, what circumstances would make you want to discard an annotation?

Patch-set: 5