[DRAFT] Add sync-c options to manifest
There are use-cases when fetching all branch is impractical and
we really need to fetch only one branch/tag.
e.g. there is a large project with binaries and every update of a
binary file is put to a separate branch.
The whole project history might be too large to allow users fetch it.

Add 'sync-c' option to 'project' and 'default' tag to make it possible
to configure 'sync-c' behavior at per-project and per-manifest level.

Change-Id: Ie36fe5737304930493740370239403986590f593