[DRAFT] Add sync_c options to manifest default and project tags
There are use cases when fetching all branch is impractical and
we really need to fetch only one branch/tag.
e.g. there is a large project with binaries and every update of a
binary file is put to a separate branch.
The whole project history might be too large to allow users fetch it.

Add "sync_c" option to 'project' tag and 'default' tag and make it possible
to configure "sync_c" behavior per-project and per-manifest level.

Change-Id: Ie36fe5737304930493740370239403986590f593