StartBranch can replace existing branch
This patch allows the StartBranch command to overwrite the HEAD of an
existing branch. This is especially useful when re-initializing the repo
manifest repository which uses a single branch (default) and needs to be
initialized properly when an existing branch named "default" exists and
the user wishes to re-initialize to a different manifest branch.

Signed-off-by: Jeremy C. Andrus <jchristian.andrus@gmail.com>

Change-Id: I69a8b55ee54b6a4d66d6ec9b41f5e9334816f4ea