Allow to use Repo without its own Git repository dependency
Allow Repo to be used without the Git ".repo/repo" for each different
repo workdir. This change also allow Repo to be fine without a ".git"
directory.
Enhance hooks symlinks with relative or absolute path depending of the
repo-wrapper path.

By default, to do not initially clone the Repo's Git, set
REPO_CLONE=False in the repo-wrapper (the "repo" file).
This allow the user to centralize Repo into a main directory of his
choice (e.g. ~/.repo/repo or /usr/share/repo). This can also be useful
for a common system installation or update.

Change-Id: I7ccc5aa0630cf49ff24bcb67d4d7688663d3b890