Update patch set 2
Patch Set 2:

I see your point.

How do you feel optional registration?  ie, we will just inherit each of the parent's attributes by default.  If an element for the submodule exists in the manifest, it can override the attributes on a one-by-one basis.

As you said in your commit message, manifest owners are at the mercy of the upstream project.  I feel like it'd be good for repo to do what it can without a manifest change everytime an upstream project adds a submodule.

Patch-set: 2
Label: Verified=0