sync: Support authentication to manifest server with .netrc
When using the --smart-sync or --smart-tag option, and the specified
manifest server is hosted on a server that requires authentication,
repo sync fails with HTTP 401 Unauthorized.

Add a new command line option, --netrc.  When this option is used,
the credentials from the .netrc file will be used to authenticate
with the manifest server using the URL syntax extension for Basic
Authentication:

  http://user:password@host:port/path

The new option may only be used in conjunction with --smart-sync or
--smart-tag, and only if the manifest server URL specified in the
manifest does not already include credentials.

Change-Id: I06e6586e8849d0cd12fa9746789e8d45d5b1f848