sync: Support authentication to manifest server
When using the --smart-sync or --smart-tag option, and the specified
manifest server is hosted on a server that requires authentication,
repo sync fails with HTTP 401 Unauthorized.

Add two new command line options, --username and --password, which
will be used to authenticate with the manifest server using the URL
syntax extension for Basic Authentication:

  http://user:password@host:port/path

The new options may only be given in conjunction with --smart-sync
or --smart-tag, and only if the manifest server URL specified in
the manifest does not already include the username and password.

Change-Id: I06e6586e8849d0cd12fa9746789e8d45d5b1f848