sync: Support authentication to manifest server with .netrc
If the user has not specified manifest server authentication
credentials with --username and --password, attempt to get
the credentials from the .netrc file.  If there is no .netrc
file in the user's home folder, or the .netrc file does not
contain settings for the manifest server, no authentication
is used.

Change-Id: I40f02c46b797cb38d4f7774746ab5b3b7a811fa4