Update patch set 1
Patch Set 1: I would prefer that you didn't submit this

(1 inline comment)

This is a good feature, but I'm not sure how I feel about all the prompting that will be going on.

If you have two remotes, you will now be prompted for 3 identities instead of 1.  There are probably use cases with even more remotes.  Not only that, but the first time it will be of the form name?/email? and the other times it will be of the form change?/name?/email?.

Another route might be that the manifest can dictate whether or not the user should be prompted for a separate identity.  Chromium knows that it needs this, Android knows it does not, maybe they should get to say so in the manifest.

There are probably some other good ways to go about this too.  Any ideas?

Marking this as -1 due to minor concerns.  Please see inline comments.

Patch-set: 1
Reviewer: Gerrit User 1006257 <1006257@173816e5-2b9a-37c3-8a2e-48639d4f1153>
Label: Code-Review=-1
Label: Verified=0