Update patch set 1
Patch Set 1:

"Here's another idea. We leave the current interaction as is and at the end say: "You will be using this identity for the following remotes: goog, aosp. If you want to setup remote specific identities, please rerun the init command with the --custom-remote-identity flag" I'm not sure what exactly that flag would look like...maybe something like --custom-remote-identity=goog (results in prompting for goog) --custom-remote-identity="goog:Bob Bobberson:bob.bobberson@google.com,aosp" (Results in prompting for aosp, sets the identity as specified for goog). The way I specified those flags is a mess, but what do you think about the general idea?"

Kind of hate it. :)

Mind you, it's a crappy problem to solve, and your approach is semi-sane/reasonable; I'm just a pain in the ass about user visible stuff like this, plus I have a bit of loathing for making people redo repo init's since for a long while, that could fail due to branch bugs; frankly just dislike that people have to re-init for every config change and that re-init'ing doesn't really take care of things fully (plus as a command, there is a difference between "re-config this one thing" and "re-initialize it all").

How about this; we add `repo config`, and include a subcommand there for setting email per remote?  Part of the reason I like this approach is that it would basically allow us to build up a set of functions for config mangling that we *know* works with a live checkout- preferably making init use that in addition.

Case in point, alternates last I knew weren't rewritten when a reference got dropped; minor, but that's code I have to maintain in chromeos's build bits; same angle, we have tools for manifest group manipulation- that's a bloody simple subcommand to add, also would give us an interface to expose to the user the various groups that exist.

Mind you that's my thoughts at 03:45, so it's entirely possible I'm being cracky; that said, thoughts?


""flaw here; the user/email isn't updated if a project switches remotes."
I'm not sure I fully understand what you were saying about forcing values, but when we sync the manifest we could watch for an added remote and suggest that the user re-run init if they want a remote-specific identity."
What I meant was if a project was remote 'cros', then moved to 'cros-internal'; the email address, unless the user has explicitly forced it manaully (aka, forced it at the repo level, not as a remote config), should auto-switch to whatever email is configured for that remote.

It's a corner case, and not one that really affect chromeos; that said for android, I could downstream/partners having this occur as they get all of their work upstreamed, and switch to working directly from the upstream repo.

Patch-set: 1