Update patch set 6
Patch Set 6: I would prefer that you didn't submit this

(1 inline comment)

I have a setup with a registered submodule and an unregistered submodule.  `repo sync` doesn't sync the unregistered submodule.  Is this intentional?

This will need to be rebased (for the .txt file).

See the inline comment as well.

Patch-set: 6
Label: Code-Review=-1
Label: Verified=0