Update patch set 2
Patch Set 2:

Hello.

I fixed the date and some other small stuff.

While I really like the overview command I'm not sure they are mergable right off the bat. 
repo info basically works in two levels. 
Level 1 gives info that is readily available. manifest, mount paths etc.
Level 2 requires an option '-a'. Which does a git fetch to compare the local with the remote.
But it does this only in the local git.

overview works on the complete repository but just on local commits.

Local commits are usually in the order of one or two per git (at least for me) but I can have a tonne of them spread out over many gits. This makes repo overview sweet.

repo infos strength on level 2 is in the remote commits where it can show you how far behind the main track you are. At least for me. 

So, to sum up, for me they work in different scopes and I'm not sure they should be merged :)

Patch-set: 2
Label: Code-Review=0