Update patch set 2
Patch Set 2:

I agree, a nice --remote option to the already available "overview" command would be great, even if it takes an option from repo config. I use overview on a daily basis (I wrote that command recently), but I definitely see value in this. I'd like to see it as an easy to access option to the overview command, because I know I will use it. I would use "info" too, but i think it already fits in nicely with the overview command (that's the idea: see an overview of what you are doing, have done, and with your changes, see what you will need to deal with soon)

Patch-set: 2