Update patch set 3
Patch Set 3:

More specifically, I'm not sure if anyone else has started using overview (people in my company do, including myself, because it helps us see how many commits we have for a project before they get merged), but since it's already there, why not just add an option to display also the opposite commits to the repo overview command?  That way there's no need to add deprecation to a command that was added 4 releases ago (v1.10.0).

That is, the current overview command (which was based on the existing "repo prune" output report) is essentially the same as doing:

 $ git log --oneline gerrit/master..HEAD

but with some added information on a per-project and per-branch basis--namely the project and branch names, the timestamp of last update, and a count of commits that are local but not yet merged.

To see the *inverse* of that (what I believe your "repo info" is trying to accomplish), would be akin to:

 $ git log --oneline HEAD..gerrit/master

It may then be out of date, but that's what the "repo sync --network-only" command is for.

Options could allow you to output the local-only commits, the remote-only commits, or both.

It's also weird that the --current-branch option was added here, but it isn't being used by the non-overview path of the command.  And the overview command uses all DAO for fetching the local branches and local commits (project.CurrentBranch, project.GetUploadableBranch(), etc), but the "info" path uses straight up git commands instead of abstracting that part out to the project.  E.g., it could use things like the project.HasChanges and ReviewableBranch.commits properties.  Could create a ReviewableBranch.remoteCommits property that does the same thing the commits property does, but reverse the heads.

I like the fact that the info command adds more information (though some things, like manifest branch, aren't really high-priority pieces of information for most people, so that too might be a good candidate for moving to an option).

Patch-set: 3