Update patch set 3
Patch Set 3:

Sorry, I think we are talking about different things.

"repo info" does not touch the network and is fast with the information it provides

"repo info -a" always touches the network and its responsetime is dependent on the network and the current git size.

"repo overview" is local yes. No network. However, it is not fast, it is not instant. This is because it goes through all the projects to find differences. Not everyone has an ssd or the whole repository in ram.

So, I am talking about response time of the commands as a whole, not network vs local. I have gits in my repo where "repo info -a" is a lot faster than repo overview even though it works over the network vs local information. As I stated in one of my first comments. It about scope. Git-scope vs repo-scope.

Patch-set: 3