Update patch set 4
Patch Set 4: I would prefer that you didn't submit this

(6 inline comments)

Anyone who is working on an even remotely fast-paced project, and might have a lot of local commits spread across different projects, is going to find this command very hard to read and hard to work with.

1. Network is a crucial downfall of this command.  Each of my projects takes in excess of 1 second to run the --all command.

2. It's visually unappealing:  by default there is no color configuration, and there is no formatting of text to make different sections stand out.  So the end result is a huge jumble of plain text that makes my head spin.

3. Not having the ability to operate against all projects at once is inconvenient at best.

TL;DR below

I'm still going to suggest that the network aspect of this be removed, because the user can always do "repo sync" first, and if I want to run the command more than once (e.g., the previous output is no longer on the screen, and I want to see it again), it's going to take in excess of 1 second for each project that I want to look at.  I just see no reason to hit the network, when all of the information is obtained from local refs anyway.  If the user *wants* to get updated information, that's exactly what the "repo sync --fetch-only" command is meant for.

It would also be nice if these lists could be formatted a bit better to see immediately the difference between sections.  The color options might actually be sufficient, but by default there is no color, so it's just a bunch of text.  If this were more visually appealing, it would give me an incentive to actually use the "all projects" model, but then I'd have to do all that manually.  As it is, I'd have to use "repo forall" or name all the projects explicitly on the command line.  All other repo commands run against all projects, unless a list of projects is given on the command line.

Patch-set: 4
Label: Code-Review=-1