Update patch set 4
Patch Set 4:

> This command is not an overview command, this command is not for getting an overview of all the commits across all the different projects. It was never meant to be that. It was meant to give information about the complete repository. Which is does.

OK, I misunderstood the purpose then.  But I also think the tasks that you're referring to are completely separate concepts, and might therefore be better suited as separate commands?  For example, to get the info printed via just "repo info" currently, it could just be a switch to the manifest command?  E.g.:

 $ repo manifest -i
 Manifest branch: master
 Manifest merge branch: refs/heads/master

Then the "repo info" command can be used to refer to specific projects and/or all projects.  This command would display the basic information about the given project(s), while still giving an *option* to work against all projects:

 $ repo info .
 Project: Development/jackd
 Mount path: /home/jhansche/sandbox/jackd
 Current revision: master
 Local Branches: 1 [SITE-6]

And if needed, you could also include information about the remote refs with a -r option (what -a does currently):

 $ repo info -r .
 Project: Development/jackd
 Mount path: /home/jhansche/sandbox/jackd
 Current revision: master
 Local Branches: 1 [SITE-6]
 Local Commits: 4 (on current branch)
 56f11576 SITE-6: commit message 4
 735bd62c SITE-6: commit message 3
 eb50ce52 SITE-6: commit message 2
 de5a0bcd SITE-6: commit message 1
 ----------------------------
 Remote Commits: 0


> The -a flag appeared when I saw the need to get a diff between the current status of my git-repo and the upstream. It was _never_ meant to give an overview of all commits in all projects. It was meant to give the status in the current git. Where it can be useful to see how far ahead the upstream is.

Now that I understand why this "info" command is useful to you and your company, I'd also like to point out that it can also be useful to others!  That is why I'm suggesting the things I am, because if it's only useful for your company, then you can use a custom repo URL and just maintain an in-house version.  But for it to be in the public version and usable by everyone, I think these are just some of the things that can make it worthwhile for *everyone*, not just Sony.


> 1) I am not sure what the argument here is. Do you want me to rename the flag to something more "networky". The functionality displayed by -a requires network access. This makes is a lot slower yes. That is why it's optional.

My point is that the -a functionality actually *doesn't* "require" network access:  it only requires network access to get an up-to-date version.  And for that, I am saying that it would be better to let the user decide when and if that network access actually occurs.  Thus the "repo sync" suggestion.

The point being that the user can still benefit from the information that is obtained from the local refs, and what your "repo info -a" command outputs -- but it might be stale.  It's no different than a "git whatchanged HEAD..gerrit/master".  If the refs/remotes/gerrit/master ref is out of date, then you know the output is also going to be out of date, and you know that "git fetch gerrit" will be required in order to get the most up-to-date information.  The same is true here:  if you *want* new information, then you can run "repo sync --network-only .", and *then* "repo info -a ." will output updated information.  The benefit is then if I want to remind myself of what was output the last time I ran it, I don't have to wait for it to hit the network yet again, even though nothing changed.

I'm just saying there's a lot of things that can already be done by other commands, and I'm not sure that combining them all into one command is the best approach:  you could create an alias for things that you want to see in one command.

The other benefit of not forcing the network usage, is that it actually *can* be useful for other people (definitely myself included!), if I *can* get the full info about all of my projects in one command (e.g., using a command like "repo info --remote --all-projects").  I usually only "repo sync" a few times a day, and yes it is slow because we have a large codebase.  So if the network usage is a requirement for the command, it will not be useful to me (and I'm sure many others), and that would be a shame, because it IS useful information to have.


> 2) Is something not working with the coloring? Have I done it wrong? You obviously see that I try to color this since you've commented on the coloring code. Please let me know if and why (if you know) it doesn't work.

Unfortunately I don't know much about how the coloring works in repo yet.  In the "overview" file, which was largely copied from "prune", my Report class contains this:

    class Report(Coloring):
      def __init__(self, config):
       	Coloring.__init__(self, config, 'status')

I noticed in your Coloring class, the 3rd argument in the __init__ call is "repoinfo".  Most likely, it's just a case of not having that "repoinfo" color configuration.  Where does that come from?  Was it something that you had to configure manually (and if so, there should probably be some defaults along with this)?  Looks like globalConfig is coming from ~/.gitconfig, so you may have an extra setting in there for:

 [color]
   repoinfo = auto

I am only looking into it now, and I may be interpreting it incorrectly.  But if you have "repoinfo = auto" in your .gitconfig, it might be just as well to just use "status" as the 3rd argument, instead of "repoinfo"?  That way it is likely to work for more people who already have the color configuration set for the most common git commands.


> repo sync --fetch-only will fetch for all projects which will take a lot longer time.

"repo sync" also has an option to fetch for only the specified projects (remember, most commands have the capability of working for 1/many/all projects).  Thus "repo sync --network-only ." will sync only the current project.  Although at that point, it's probably easier to just run "git fetch" instead.


> My current working repo has 522 projects in it. The reason to hit the network is to get the diff. To know how far behind the upstream you are. If you do not find that useful, do not use it.

I understand that, but I still think it's a bad thing to force the network access on the user, without letting him decide.  The user will *know* that the information will be out of date until the next repo sync.  I don't think that's a bad thing, and it is in line with the functionality of both repo and even git itself.

Patch-set: 4