Update patch set 5
Patch Set 5:

"""This commit only change the Repo behavior when the command is not in the parents directory, but elsewhere. This will change nothing for the current use."""

I recognize that, but it adds code complexity for the sake of an unsupported use case.

"""About the self update thing, it's only from the commit I2f7e49cc. Again, there is no Repo behavior change for the current use. The commits only affect installation without the ".git" directory, which happen only in a manual way (i.e. git archive). Anyway, without this commit, Repo will crash if there is no ".git" directory, so it's more a fix than feature ;)"""

Same comment here.  Currently, there is a design assumption that repo should always be able to update itself regardless of where it is installed.

"""Theses commits are necessary before being able to package Repo."""

I'm not sure that being packageable is that desirable of a feature.

I appreciate the work you've put into these changes, but I suggest that you start a thread on repo-discuss@googlegroups.com and see if others in the community would like to see these use cases supported and if other repo authors are willing to abandon the design assumption that repo will self-update.

Patch-set: 5