sync: Negotiate with persistent-https proxy only once
The overhead of negotiating with the git-remote-persistent-https proxy
process with a high -j can slow down sync significantly, especially on
older machines. Assuming the remote in all projects uses the same
protocol (persistent-http or persistent-https), this work is wasted,
since the proxy configuration is identical for each request.

Instead of negotiating on every fetch, teach sync to parse the proxy
config outputted by the -print_config flag and pass this down to
Sync_NetworkHalf. This also requires an ugly insteadof dance to make
sure git doesn't insteadof the http(s) URL back to persistent-http(s),
since the documented way to enable the persistent-https proxy is to do
a global url.persistent-https://.insteadof=https://.

Change-Id: I47e566fd7c956f07b63f1951d85e52a7124630cf