Update patch set 1
Patch Set 1:

Shawn, this is so you can see if there's anything dumb I'm doing, and so there's a stable version on a server.

This depends on a flag to git-remote-persistent-https that only exists in my local copy, so if anybody else outside Google is interested in trying this out, let me know and I'll make that patch available.

Patch-set: 1
Label: Code-Review=0